requirements.txt
-e git+https://github.com/sumanurawat/ai_agent_utilities.git@7bed3d389f090822c71bd5d7ee64ea3baa157262#egg=ai_agent_utilities
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dateparser==1.2.1
distro==1.9.0
filelock==3.18.0
fonttools==4.56.0
fpdf==1.7.2
frozendict==2.4.6
GoogleNews==1.6.15
griffe==1.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.8
lxml==5.3.1
matplotlib==3.10.1
multitasking==0.0.11
nltk==3.9.1
numpy==2.2.3
openai==1.66.1
openai-agents==0.0.2
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
praw==7.8.1
prawcore==2.4.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
snscrape==0.7.0.20230622
soupsieve==2.6
textblob==0.19.0
tqdm==4.67.1
types-requests==2.32.0.20250306
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
update-checker==0.18.0
urllib3==2.3.0
websocket-client==1.8.0
wordcloud==1.9.4
yfinance==0.2.54
